Agregue ResponseFactory y validaciones a los EndPoint de User

+++ Controllers/UserController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using TpIntegradorSofttek.DTOs;
 using TpIntegradorSofttek.Entities;
 using TpIntegradorSofttek.Helper;
+using TpIntegradorSofttek.Infrastructure;
 using TpIntegradorSofttek.Services;
 
 namespace TpIntegradorSofttek.Controllers
@@ -22,22 +23,22 @@ namespace TpIntegradorSofttek.Controllers
 
         [HttpGet("GetAllActive")]
         [Authorize]
-        public async Task<ActionResult<IEnumerable<User>>> GetAllActive()
+        public async Task<IActionResult> GetAllActive()
         {
             var users = await _unitOfWork.UserRepository.GetAllActive();
 
-            return Ok(users);
+            return ResponseFactory.CreateSuccessResponse(200, users);
         }
 
 
         //devuelve el usuario incluso si esta dado de baja (logicamente)
         [HttpGet("GetById/{id}")]
         [Authorize]
-        public async Task<ActionResult<User>> GetById([FromRoute] int id)
+        public async Task<IActionResult> GetById([FromRoute] int id)
         {
             var users = await _unitOfWork.UserRepository.GetById(new User(id));
 
-            return Ok(users);
+            return ResponseFactory.CreateSuccessResponse(200, users);
         }
 
 
@@ -46,29 +47,62 @@ namespace TpIntegradorSofttek.Controllers
         [Authorize(Policy = "Admin")]
         public async Task<IActionResult> Register(RegisterDto dto)
         {
+            if (await _unitOfWork.UserRepository.UserExByMail(dto.Email)) return ResponseFactory.CreateErrorResponse(409, $"Ya existe un usuario registrado con el mail: {dto.Email}");
             var user = new User(dto);
             await _unitOfWork.UserRepository.Insert(user);
             await _unitOfWork.Complete();
-            return Ok(true);
+
+
+            return ResponseFactory.CreateSuccessResponse(201,"Usuario registrado con exito!");
         }
 
+
+        //antes de actualizar verifico si el usuario ya existe un usario con el mail a actualizar
+        //si ya existe verifico que ese usuario con el mail existente sea el mismo usuario que se esta actualizando
+        //si no lo es no lo dejo actualizar.
         [HttpPut("Update/{id}")]
         [Authorize(Policy = "Admin")]
         public async Task<IActionResult> Update([FromRoute] int id, RegisterDto dto)
         {
+            if (await _unitOfWork.UserRepository.UserExByMail(dto.Email))
+            {
+                var existingUser = await _unitOfWork.UserRepository.GetByEmail(dto.Email);
+
+                if (existingUser.CodUser != id)
+                {
+                    return ResponseFactory.CreateErrorResponse(409, $"Ya existe un usuario registrado con el mail: {dto.Email}");
+                }
+            }
             var result = await _unitOfWork.UserRepository.Update(new User(dto, id));
             await _unitOfWork.Complete();
-            return Ok(true);
+            return ResponseFactory.CreateSuccessResponse(201, "Usuario actualizado con exito!");
         }
 
+
+        //verifica que exista un usuario con el id ingresado, si existe verifica que este activo, si no esta activo indica error 409
+        //si no existe usuario con id ingresado indica error 404
         [HttpDelete("Delete/{id}")]
         [Authorize(Policy = "Admin")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
-            var result = await _unitOfWork.UserRepository.Delete(new User(id));
-            await _unitOfWork.Complete();
-            return Ok(true);
+            if (await _unitOfWork.UserRepository.UserExById(id))
+            {
+                var user = await _unitOfWork.UserRepository.GetById(new User(id));
+                if (user.IsActive)
+                {
+                    var result = await _unitOfWork.UserRepository.Delete(new User(id));
+                    await _unitOfWork.Complete();
+                    return ResponseFactory.CreateSuccessResponse(201, "Usuario eliminado con exito!");
+                }
+                else
+                {
+                    return ResponseFactory.CreateErrorResponse(409, $"El Usuario con Id: {id} ya se encuentra eliminado");
+                }
+            }
+            return ResponseFactory.CreateErrorResponse(404, $"No existe ningun usuario con el Id: {id}");
         }
+
+        
 
 
     }
